Remove duplicates skeletons
