+++ app/issues/[id]/DeleteIssueButton.tsx
@@ -14,7 +14,7 @@ const DeleteIssueButton = ({ issueId }: { issueId: number }) => {
         try {
             setDeleting(true);
             await axios.delete('/api/issues/' + issueId);
-            router.push('/issues');
+            router.push('/issues/list');
             router.refresh();
         } catch (error) {
             setDeleting(false);
