+++ app/issues/_components/IssueForm.tsx
@@ -38,7 +38,7 @@ const IssueForm = ({issue}:{issue?:Issue}) => {
         await axios.patch('/api/issues/' + issue.id, data);
       else
         await axios.post('/api/issues', data)
-      router.push('/issues');
+      router.push('/issues/list');
       router.refresh();
     } catch (error) {
       setSubmitting(false);
